pyproject.toml
[tool.poetry]
name = "dj_sinp_nomenclatures"
version = "0.1.0"
description = "Django app to manage french SINP nomenclatures"
authors = ["dbChiro project <project@dbchiro.org>"]
license = "AGPLv3"
packages = [
    { include = "sinp_nomenclatures" },
]


[tool.poetry.dependencies]
python = "^3.6.2"
Django = "^3.2.5"
django-rest-framework = "^0.1.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
